Fix recurrent evaluation
To evaluate a recurrent neural network new inputs and outputs are
introduced by the `unroll` operation to account for and ultimately
compute the recurrent connections. Internally the recurrent fabricator
just unrolls and then delegates to the feedforward fabricator. The
feedforward fabricator does not know or care about the originally
recurrent structure and treats all inputs and outputs the same. There a
problem arises: To guarantee that the outputs are presented in the same
order every time the net is constructed, they are sorted. The recurrent
evaluator then assumes that "the first x, x being th configured amount
of outputs by the user" correspond to the network output. Due to the
sorting this does not always hold true and eventually outputs that just
represent internal state will be read and returned. Further new outputs
could arise over the course of evolution once again making it impossible
to "just sort" the outputs.

The feedworward evaluator should not need to care an just sort as it
does. The important feature here is that the same output node appears on
the same position in the output vector for every fabricated net every
time.

This is broken by dynamically introducing new output nodes which breaks
the assumption that sorting guarantees the above statement. The
introduction of nodes needs to be handeled more carefully. As the entire
"breakage" occurs inside the recurrent fabricator it should account for
the actually desired behavior.

Now ids static inputs and outputs are remapped in the network so that
sorting again is meaningful in that it always sorts the static output
nodes into the first x positions of the output vector. That requires
remapping all connections that originate from or run into nodes with
remapped ids. Further during remapping one must not introduce collisions
, i.e. mapping to an already existing id. The implementation of this is
inside the unroll operation.