Fix node ordering in feedforward evaluation.
The crate was build with the assumption that nodes could always be
meaningfully ordered by theit ids. This was a leaky abstraction from the
using crate, namely set-genome which initially incorporated numerically
ascending ids so the assumption did hold. A while back the create
transitioned to hashes as ids breaking this assumption and introducing
buggy behavior in this crate. Now they are sorted according to the
Ord trait implementation, which is a reasonable as it communicates the
possibility to order stuff. A reasonable implementation of Ord still
lies with the external crates.