Fix SparseMatrixRecurrentEvaluator
As the maximum row index + 1 was taken as the dimension of the rows it
could produce dimension missmatches when constructing stages in cases
where the highest input dimension was unused for computing the next
stage. This has been fixed now by using the dimensionality of the input
(length of available_nodes vector) as the proper dimension.